% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydricsoils.R
\docType{data}
\name{fihs}
\alias{fihs}
\title{DATA: 'Field Indicators of Hydric Soils in the United States'}
\format{
An object of class \code{data.frame} with 45 rows and 11 columns.
}
\usage{
data(fihs)
}
\description{
This dataset is derived from the latest edition of the \href{https://www.nrcs.usda.gov/sites/default/files/2024-09/Field-Indicators-of-Hydric-Soils.pdf}{'Field Indicators of Hydric Soils in the United States' (Version 9.0, 2024)}
}
\references{
United States Department of Agriculture, Natural Resources Conservation Service. 2018. Field Indicators of Hydric Soils in the United States, Version 8.2. L.M. Vasilas, G.W. Hurt, and J.F. Berkowitz (eds.). USDA, NRCS, in cooperation with the National Technical Committee for Hydric Soils. Available online: \url{https://www.nrcs.usda.gov/resources/guides-and-instructions/field-indicators-of-hydric-soils}
}
\keyword{datasets}
