% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydricsoils.R
\name{hydricsoils_data_dir}
\alias{hydricsoils_data_dir}
\title{hydricsoils Data Directory}
\usage{
hydricsoils_data_dir()
}
\value{
\code{hydricsoils_data_dir()}: character. Path to hydricsoil package user
data directory. Default: \code{tools::R_user_dir("hydricsoils", which = "data")}
}
\description{
hydricsoils Data Directory
}
\examples{

hydricsoils_data_dir()

}
