% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator-conversion.R
\name{indicator_to_name}
\alias{indicator_to_name}
\alias{name_to_indicator}
\alias{indicator_to_usesym}
\alias{usesym_match}
\alias{usesym_to_indicator}
\alias{fihs_match}
\title{Convert between Indicator Codes, Names and Area Symbols}
\usage{
indicator_to_name(x)

name_to_indicator(x)

indicator_to_usesym(x, simplify = TRUE)

usesym_match(x)

usesym_to_indicator(x, simplify = TRUE)

fihs_match(x, a)
}
\arguments{
\item{x}{character. Vector of indicator codes or names.}

\item{simplify}{\emph{logical}. If result is a \emph{list}, and \code{x} is length \code{1}, return a character vector. Default: \code{TRUE}}

\item{a}{character. Used only for \code{fihs_match()}. Column name in \code{fihs} to match \code{x} against.}
}
\value{
character or list. Vector of indicator codes or names corresponding to \code{x}.
}
\description{
Helper functions for converting between symbols and names associated with hydric indicators, their names, and Land Resource Regions (LRR) or Major Land Resource Areas (MLRA) using the internal dataset \code{lrrmlra}.
}
\examples{

indicator_to_name(c("A11", "A10", NA, "F1", "A8"))

name_to_indicator(c("Depleted Below Dark Surface", "2 cm Muck", NA, "Loamy Mucky Mineral", "Muck Presence"))

indicator_to_usesym(c("A11", "A10", NA, "F1", "A8"))

usesym_to_indicator(c("22A", "150A"))
}
\references{
United States Department of Agriculture, Natural Resources Conservation Service. 2018. Field Indicators of Hydric Soils in the United States, Version 8.2. L.M. Vasilas, G.W. Hurt, and J.F. Berkowitz (eds.). USDA, NRCS, in cooperation with the National Technical Committee for Hydric Soils. Available online: \url{https://www.nrcs.usda.gov/resources/guides-and-instructions/field-indicators-of-hydric-soils}

United States Department of Agriculture, Natural Resources Conservation Service. 2022. Land resource regions and major land resource areas of the United States, the Caribbean, and the Pacific Basin. U.S. Department of Agriculture, Agriculture Handbook 296. Available online: \url{https://www.nrcs.usda.gov/resources/data-and-reports/major-land-resource-area-mlra}
}
\seealso{
\link{fihs}
}
