% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_lrrmlra_geometry.R
\name{cache_lrrmlra_geometry}
\alias{cache_lrrmlra_geometry}
\alias{clear_lrrmlra_geometry}
\alias{lrrmlra_geometry}
\title{Download and Cache or Load Latest 'MLRA Geographic Database'}
\source{
The default source is a ZIP file containing an ESRI Shapefile. The direct download URL is \url{https://www.nrcs.usda.gov/sites/default/files/2022-10/MLRA_52_2022.zip}, which can be found \href{https://www.nrcs.usda.gov/resources/data-and-reports/major-land-resource-area-mlra}{here}.
}
\usage{
cache_lrrmlra_geometry(
  overwrite = FALSE,
  dsn =
    "/vsizip//vsicurl/https://www.nrcs.usda.gov/sites/default/files/2022-10/MLRA_52_2022.zip/MLRA_52_2022"
)

clear_lrrmlra_geometry()

lrrmlra_geometry(
  overwrite = FALSE,
  dsn =
    "/vsizip//vsicurl/https://www.nrcs.usda.gov/sites/default/files/2022-10/MLRA_52_2022.zip/MLRA_52_2022"
)
}
\arguments{
\item{overwrite}{Overwrite existing file? Passed to \code{terra::writeVector()}. Default: \code{FALSE}}

\item{dsn}{Path to a geospatial data source. Default path uses GDAL Virtual File Systems to access and decompress remote source (\code{"/vsizip//vsicurl/https://www.nrcs.usda.gov/sites/default/files/2022-10/MLRA_52_2022.zip/MLRA_52_2022"}).}
}
\value{
\code{cache_lrrmlra_geometry()}: logical. Called for the side-effect of adding the MLRA geometry file to the cache. Returns \code{TRUE} if the \code{"lrrmlra.gpkg"} file exists in the user's \code{"data"} cache directory.

\code{clear_lrrmlra_geometry()}: logical. Called for the side-effect of removing the MLRA geometry file from the cache. Returns \code{TRUE} if \code{"lrrmlra.gpkg"} is successfully removed from user data cache.

\code{lrrmlra_geometry()}: A terra \emph{SpatVector} object containing \code{lrrmlra} attributes and geometry.
}
\description{
Creates a local copy of the latest 'MLRA Geographic Database' in the user's package data cache. This file/directory is separate from the package installation.
}
\examples{
\dontrun{
 library(terra)

 # download (if needed) and load geometry
 x <- lrrmlra_geometry()
 plot(x)

 # clear cache
 clear_lrrmlra_geometry()

 # download and cache (without loading)
 cache_lrrmlra_geometry()
}

}
\references{
United States Department of Agriculture, Natural Resources Conservation Service. 2022. Land resource regions and major land resource areas of the United States, the Caribbean, and the Pacific Basin. U.S. Department of Agriculture, Agriculture Handbook 296. Available online: \url{https://www.nrcs.usda.gov/resources/data-and-reports/major-land-resource-area-mlra}
}
\seealso{
\link{lrrmlra}
}
