% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrr-mlra-conversion.R
\name{mlra_to_lrr}
\alias{mlra_to_lrr}
\alias{mlra_to_mlraname}
\alias{mlraname_to_mlra}
\alias{lrr_to_mlra}
\alias{lrr_to_lrrname}
\alias{lrrname_to_lrr}
\title{Convert between Land Resource Region (LRR) or Major Land Resource Area (MLRA) Symbols and Names}
\usage{
mlra_to_lrr(x)

mlra_to_mlraname(x)

mlraname_to_mlra(x)

lrr_to_mlra(x)

lrr_to_lrrname(x)

lrrname_to_lrr(x)
}
\arguments{
\item{x}{character. Vector LRR or MLRA symbols or names.}
}
\value{
character. Vector of LRR or MLRA symbols or names corresponding to \code{x}.
}
\description{
Helper functions for converting between symbols and names associated with Land Resource Regions (LRR) or Major Land Resource Areas (MLRA) using the internal dataset \code{lrrmlra}.
}
\examples{

mlra_to_lrr(c("18", "22A"))

mlra_to_mlraname("18")

mlraname_to_mlra("Sierra Nevada and Tehachapi Mountains")

lrr_to_mlra("C")

lrr_to_lrrname("D")

lrrname_to_lrr("California Subtropical Fruit, Truck, and Specialty Crop Region")

}
\references{
United States Department of Agriculture, Natural Resources Conservation Service. 2022. Land resource regions and major land resource areas of the United States, the Caribbean, and the Pacific Basin. U.S. Department of Agriculture, Agriculture Handbook 296. Available online: \url{https://www.nrcs.usda.gov/resources/data-and-reports/major-land-resource-area-mlra}
}
\seealso{
\link{lrrmlra}
}
