% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydricsoils.R
\docType{data}
\name{lrrmlra}
\alias{lrrmlra}
\title{DATA: 'Land Resource Regions and Major Land Resource Areas of the United States, the Caribbean, and the Pacific Basin'}
\format{
An object of class \code{data.frame} with 267 rows and 5 columns.
}
\usage{
data(lrrmlra)
}
\description{
This dataset is derived from the latest edition of the \href{https://www.nrcs.usda.gov/sites/default/files/2022-10/MLRA_52_2022.zip}{'MLRA Geographic Database' (Version 5.2, 2022)}
}
\references{
United States Department of Agriculture, Natural Resources Conservation Service. 2022. Land resource regions and major land resource areas of the United States, the Caribbean, and the Pacific Basin. U.S. Department of Agriculture, Agriculture Handbook 296. Available online: \url{https://www.nrcs.usda.gov/resources/data-and-reports/major-land-resource-area-mlra}
}
\keyword{datasets}
