% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_usage.R
\name{plot_indicator_usage}
\alias{plot_indicator_usage}
\title{Plot the Usage Area of a Field Indicator for Hydric Soils}
\usage{
plot_indicator_usage(
  indicator,
  ext = NULL,
  plg = NULL,
  crs = NULL,
  test_areas = FALSE,
  overwrite = FALSE,
  dsn = NULL
)
}
\arguments{
\item{indicator}{\emph{character}. The alphanumeric code corresponding to the indicator of interest. For example: \code{"A12"}}

\item{ext}{\emph{SpatExtent}, \emph{SpatVector}, \emph{SpatRaster}, or \emph{character}. Either a terra object that can be converted to SpatExtent. Alternately, a \emph{character} vector denoting the usage constraints (one or more of \code{"Approved"}, \code{"Excluded"}, or \code{"Testing"}). The resulting \emph{SpatExtent} is passed as the \code{ext} argument of \code{terra::plot()}.}

\item{plg}{\emph{list}. Passed as \code{plg} argument of \code{terra::plot()}.}

\item{crs}{\emph{character}. Optional: target coordinate reference system to project \code{lrrmlra_geometry()} result to. Default: \code{NULL} uses the \code{"OGC:CRS84"} geographic coordinate system.}

\item{test_areas}{\emph{logical}. Show areas where indicator is being tested? Default: \code{FALSE}}

\item{overwrite}{\emph{logical}. Passed to \code{lrrmlra_geometry()}. Default: \code{FALSE}}

\item{dsn}{\emph{character}. Passed to \code{lrrmlra_geometry()}. Default: \code{NULL} uses the default remote source via GDAL Virtual File System.}
}
\value{
This function is called for the side effect of creating output on the graphics pane.
}
\description{
Plot the Usage Area of a Field Indicator for Hydric Soils
}
\examples{
\dontshow{if (!inherits(try(requireNamespace("terra", quietly = TRUE), silent = TRUE), 'try-error')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

par(mfrow = c(2, 1))
plot_indicator_usage("A5", test_areas = TRUE, ext = c("Approved", "Excluded"))
plot_indicator_usage("A9", test_areas = TRUE, ext = c("Approved", "Excluded"))
\dontshow{\}) # examplesIf}
}
