% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydricsoils-package.R
\docType{package}
\name{hydricsoils-package}
\alias{hydricsoils}
\alias{hydricsoils-package}
\title{hydricsoils: Tools for working with 'Field Indicators of Hydric Soils' and related data}
\description{
The goal of 'hydricsoils' is to provide easy access to definitions, criteria, and area of applicability for 'Field Indicators of Hydric Soils in the United States' and also to provide tools which assist in evaluating associated soil morphology.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://humus.rocks/hydricsoils/}
}

}
\keyword{internal}
